requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# Logging
loguru==0.7.2

# UUID v7
uuid6==2024.1.12

# JSON Serialization
orjson==3.9.15

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Email validation
email-validator==2.1.0

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
ruff==0.1.14
mypy==1.8.0
